updated process.h & process.cpp file to add const func for pids.
This was necessary because in the '<' operator overloading in
process.cpp file, const `process` object has been passed which
means that if that object needs to be compared based on the
CPU Utilization of a process then that utilization function
also needs to be const. In that case, all the required supporting
function should also be const. So, one of the required function
is to get the pid for the path in "/proc/[pid]/stat". Default
int Pid() function is non-const. So, it won't work.